training/pyproject.toml
[project]
name = "fever-dream-training"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.0",
    "torchvision>=0.15",
    "numpy",
    "Pillow",
    "opencv-python",
    "setuptools<81",
    "tensorboard>=2.0",
    "onnx>=1.14",
    "onnxruntime>=1.15",
    "tqdm",
]
